Provide a Dockerfile and instructions to run the notebooks using it.
The existing installation instructions for building Gen locally are complex
(and contain some errors). I anticipate that users will face difficulty getting
those instructions to work on their machien so a Dockerized system should be easier
for users who just need a "Quick Installation".

In the next commit I will populate the "Advanced Installation" section with a
documentation of the steps taken in the Dockerfile that address some issues with
the current documentation.