Simplify README.md to refer to Dockerfile.
The commands for building IJulia with the JUPYTER flag
or PyCall/PyPlot with the ENV["PYTHON"] configuration are not
necessary. On Linux, Julia obtains these variables from the PATH
which is what the existing commands did.

Natively building on Ubuntu 16.04 should be trivial by reading
the Dockerfile. Maybe we should provide instructions for building
on a Mac OS X.